chore: refactored layouts
