+++ src/app/[locale]/profile/basic-info/page.tsx
@@ -1,0 +1,15 @@
+import { useTranslations } from "next-intl";
+import { BasicInfoForm } from "./form";
+
+export default function ProfileBasicInfoPage() {
+  const t = useTranslations("organization.basic_info_page");
+
+  return (
+    <>
+      <h1 className="heading">{t("heading")}</h1>
+      <section>
+        <BasicInfoForm />
+      </section>
+    </>
+  );
+}
