+++ src/layout/page/index.ts
@@ -1,0 +1,1 @@
+export { ProfileLayout } from "./ProfileLayout";
